* file architecture refactore
