+++ Sources/WishKit/Shared/Model/LocalWishState.swift
