+++ Sources/WishKit/Shared/Validation/WishValidationEmailValidationResult.swift
