+++ Sources/WishKit/Shared/ViewModel/DetailWishViewModel.swift
