+++ Sources/WishKit/Shared/ViewModel/WishlistViewModel.swift
