+++ Sources/WishKit/macOS/Extensions/View+ToolbarKeyboardDoneButton+macOS.swift
@@ -1,0 +1,10 @@
+#if os(macOS)
+import SwiftUI
+
+extension View {
+    @ViewBuilder
+    func toolbarKeyboardDoneButton() -> some View {
+        self
+    }
+}
+#endif
