+++ Sources/WishKit/visionOS/Extensions/Bundle+AppIcon+visionOS.swift
@@ -1,0 +1,18 @@
+#if os(visionOS)
+import UIKit
+
+extension Bundle {
+    public var appIcon: UIImage? {
+        if
+            let appIcons = infoDictionary?["CFBundleIcons"] as? [String: Any],
+            let primaryAppIcon = appIcons["CFBundlePrimaryIcon"] as? [String: Any],
+            let appIconFiles = primaryAppIcon["CFBundleIconFiles"] as? [String],
+            let lastAppIcon = appIconFiles.last
+        {
+            return UIImage(named: lastAppIcon)
+        }
+
+        return nil
+    }
+}
+#endif
