+++ Sources/WishKit/visionOS/Extensions/View+ToolbarKeyboardDoneButton+visionOS.swift
@@ -1,0 +1,10 @@
+#if os(visionOS)
+import SwiftUI
+
+extension View {
+    @ViewBuilder
+    func toolbarKeyboardDoneButton() -> some View {
+        self
+    }
+}
+#endif
