+++ Sources/WishKit/visionOS/Wishlist/View/WishlistContainer+visionOS.swift
@@ -1,0 +1,129 @@
+#if os(visionOS)
+import SwiftUI
+import WishKitShared
+
+struct WishlistContainer: View {
+
+    @Environment(\.colorScheme)
+    private var colorScheme
+
+    @StateObject
+    private var viewModel = WishlistViewModel()
+
+    @State
+    private var isRefreshing = false
+
+    @State
+    private var showingCreateSheet = false
+
+    @ObservedObject
+    private var wishModel: WishModel
+
+    init(wishModel: WishModel) {
+        self.wishModel = wishModel
+        self.wishModel.fetchList()
+    }
+
+    func refreshList() {
+        isRefreshing = true
+        wishModel.fetchList {
+            isRefreshing = false
+        }
+    }
+
+    var body: some View {
+        VStack {
+            switch WishKit.config.buttons.segmentedControl.display {
+            case .show:
+                segmentedControlView
+            case .hide:
+                noSegmentedControlView
+            }
+
+            WishlistView(
+                wishModel: wishModel,
+                selectedWishState: $viewModel.selectedWishState
+            )
+                .background(systemBackgroundColor)
+
+        }.background(systemBackgroundColor)
+    }
+
+    var segmentedControlView: some View {
+        HStack {
+
+            if WishKit.config.buttons.segmentedControl.display == .show {
+                Picker("", selection: $viewModel.selectedWishState) {
+                    ForEach(viewModel.feedbackStateSelection, id: \.self) { state in
+                        Text("\(state.description) (\(viewModel.count(for: state, wishModel: wishModel)))")
+                            .tag(state)
+                    }
+                }.frame(maxWidth: 150)
+            }
+
+            Spacer()
+
+            Button(action: refreshList) {
+                if isRefreshing {
+                    Text(WishKit.config.localization.refreshing)
+                } else {
+                    Text(WishKit.config.localization.refresh)
+                }
+            }
+
+            if WishKit.config.buttons.addButton.location == .navigationBar {
+                Button(action: { self.showingCreateSheet.toggle() }) {
+                    Text(WishKit.config.localization.addButtonInNavigationBar)
+                }
+                .padding(.leading, 15)
+                .sheet(isPresented: $showingCreateSheet) {
+                    CreateWishView(
+                        createActionCompletion: { wishModel.fetchList() },
+                        closeAction: { self.showingCreateSheet = false }
+                    )
+                    .frame(minWidth: 500, idealWidth: 500, minHeight: 400, maxHeight: 600)
+                    .background(systemBackgroundColor)
+                }
+            }
+        }.padding(EdgeInsets(top: 15, leading: 15, bottom: 15, trailing: 20))
+    }
+
+    var noSegmentedControlView: some View {
+        HStack {
+            Button(action: refreshList) {
+                if isRefreshing {
+                    Text(WishKit.config.localization.refreshing)
+                } else {
+                    Text(WishKit.config.localization.refresh)
+                }
+            }.padding(.trailing, WishKit.config.buttons.addButton.location == .floating ? 15 : 0)
+
+            if WishKit.config.buttons.addButton.location == .navigationBar {
+                Button(action: { self.showingCreateSheet.toggle() }) {
+                    Text(WishKit.config.localization.addButtonInNavigationBar)
+                }
+                .padding(.leading, 15)
+                .sheet(isPresented: $showingCreateSheet) {
+                    CreateWishView(
+                        createActionCompletion: { wishModel.fetchList() },
+                        closeAction: { self.showingCreateSheet = false }
+                    )
+                    .frame(minWidth: 500, idealWidth: 500, minHeight: 400, maxHeight: 600)
+                    .background(systemBackgroundColor)
+                }
+            }
+        }.padding(EdgeInsets(top: 15, leading: 5, bottom: 15, trailing: 15))
+    }
+
+    var systemBackgroundColor: Color {
+        switch colorScheme {
+        case .light:
+            WishKit.theme.tertiaryColor?.light ?? PrivateTheme.systemBackgroundColor.light
+        case .dark:
+            WishKit.theme.tertiaryColor?.dark ?? PrivateTheme.systemBackgroundColor.dark
+        @unknown default:
+            WishKit.theme.tertiaryColor?.light ?? PrivateTheme.systemBackgroundColor.light
+        }
+    }
+}
+#endif
